refactoring - 리뷰반영: 테스트패키지 변경
