+++ src/test/java/nextstep/subway/unit/LineSectionsUnitTest.java
@@ -1,6 +1,8 @@
-package nextstep.subway.section;
+package nextstep.subway.unit;
 
 import nextstep.subway.line.Line;
+import nextstep.subway.section.Section;
+import nextstep.subway.section.Sections;
 import nextstep.subway.station.Station;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
